+++ src/test/java/chess/domain/positionFilter/PawnTest.java
@@ -7,12 +7,12 @@ import chess.domain.board.position.Direction;
 import chess.domain.board.position.Position;
 import chess.domain.board.position.Row;
 import chess.domain.game.PositionsFilter;
-import chess.domain.mock.BlackPieceRepository;
-import chess.domain.mock.NotExistsPieceRepository;
-import chess.domain.mock.WhitePieceRepository;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
+import chess.repository.fake.BlackPieceRepository;
+import chess.repository.fake.NotExistsPieceRepository;
+import chess.repository.fake.WhitePieceRepository;
 import java.util.List;
 import java.util.Map;
 import java.util.Queue;
